requirements.txt
asgiref==3.11.0
Django==4.1.13
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.3.6
psycopg-binary==3.3.2
PyJWT==2.10.1
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.2.4
tzdata==2025.3
